Move insert_use util to utils
